nixos/overoptimized-recursive-importer: init
I couldn't bring myself to commit this to my config. It is actually
faster - I get a 40ms improvement in nixpkgs when using this version.
But in a normal config, you're going to get something closer to 2ms. I
can't justify the loss to readability, especially for a function I'm
recommending people to steal, use, and modify themselves.